fix: separated editor constants, graphql editor css, pagination isSeparate flag
