+++ app/client/src/pages/Editor/APIEditor/CommonEditorForm.tsx
@@ -3,8 +3,8 @@ import { useDispatch, useSelector } from "react-redux";
 import {
   HTTP_METHOD_OPTIONS,
   API_EDITOR_TABS,
-  GRAPHQL_HTTP_METHOD_OPTIONS,
-} from "constants/ApiEditorConstants";
+} from "constants/ApiEditorConstants/CommonConstants";
+import { GRAPHQL_HTTP_METHOD_OPTIONS } from "constants/ApiEditorConstants/GraphQLEditorConstants";
 import styled from "styled-components";
 import FormLabel from "components/editorComponents/FormLabel";
 import FormRow from "components/editorComponents/FormRow";
