+++ app/client/src/pages/Editor/APIEditor/GraphQL/QueryWrapperWithCSS.tsx
@@ -1,3 +1,4 @@
+import { Colors } from "constants/Colors";
 import styled from "styled-components";
 
 const QueryWrapper = styled.div`
@@ -238,6 +239,14 @@ const QueryWrapper = styled.div`
   .CodeMirror-jump-token {
     text-decoration: underline;
     cursor: pointer;
+  }
+
+  &&&&& .CodeMirror {
+    border: 0;
+  }
+
+  &&& .CodeMirror-gutters {
+    background: ${Colors.GRAY_50};
   }
 `;
