+++ app/client/src/pages/Editor/APIEditor/RapidApiEditorForm.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { connect } from "react-redux";
 import { reduxForm, InjectedFormProps, formValueSelector } from "redux-form";
-import { POST_BODY_FORMAT_OPTIONS } from "constants/ApiEditorConstants";
+import { POST_BODY_FORMAT_OPTIONS } from "constants/ApiEditorConstants/CommonConstants";
 import styled from "styled-components";
 import FormLabel from "components/editorComponents/FormLabel";
 import FormRow from "components/editorComponents/FormRow";
