+++ app/client/src/pages/Editor/Explorer/ExplorerIcons.tsx
@@ -4,7 +4,10 @@ import { Colors } from "constants/Colors";
 import { Plugin } from "api/PluginApi";
 import ImageAlt from "assets/images/placeholder-image.svg";
 import styled from "styled-components";
-import { HTTP_METHOD, HTTP_METHODS_COLOR } from "constants/ApiEditorConstants";
+import {
+  HTTP_METHOD,
+  HTTP_METHODS_COLOR,
+} from "constants/ApiEditorConstants/CommonConstants";
 import { PRIMARY_KEY, FOREIGN_KEY } from "constants/DatasourceEditorConstants";
 import { Icon } from "@blueprintjs/core";
 import { ControlIcons } from "icons/ControlIcons";
